aider: Added custom password reset template with warm, understanding message and link to reset page.
# Aider chat conversation:

USER: git add `backend\welcome.py`

USER: let's add a custom password reset template. Add a warm, understanding message for the user. This link is accessible from the login page. In the case the user has entered an email address, pass that to the reset page so they don't have to enter it again

ASSISTANT: To add a custom password reset template, we need to create a new HTML file in the `app/templates/security` directory. Let's call this file `reset_password.html`.

In this file, we will create a form that will allow the user to enter their email address and submit it to reset their password. We will also add a warm, understanding message for the user.

Here is the edit block for the new file:

```html
app/templates/security/reset_password.html
<<<<<<< HEAD

=======
<!DOCTYPE html>
<html>
<head>
    <title>Reset Password</title>
</head>
<body>
    <h1>Reset Password</h1>
    <p>We understand that things happen. Let's get you a new password.</p>
    <form method="POST" action="{{ url_for_security('reset_password') }}">
        <input type="email" name="email" placeholder="Email" value="{{ email }}" required>
        <input type="submit" value="Reset Password">
    </form>
</body>
</html>
>>>>>>> updated
```

Next, we need to modify the `login_user.html` file to include a link to the password reset page. The link should be placed near the password input field.

Here is the edit block for the `login_user.html` file:

```html
app/templates/security/login_user.html
<<<<<<< HEAD
    {{ login_user_form.password.label }}<br>
    {{ login_user_form.password(size=32, placeholder="Enter your desired password") }}
    {% for error in login_user_form.password.errors %}
        <span class="form-error">{{ error }}</span>
    {% endfor %}
    <br>
=======
    {{ login_user_form.password.label }}<br>
    {{ login_user_form.password(size=32, placeholder="Enter your desired password") }}
    {% for error in login_user_form.password.errors %}
        <span class="form-error">{{ error }}</span>
    {% endfor %}
    <br>
    <a href="{{ url_for_security('reset_password', email=login_user_form.email.data) }}">Forgot password?</a>
    <br>
>>>>>>> updated
```

Please replace `url_for_security('reset_password')` with the actual URL for the password reset page in your application.